pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wagtail-sitemap-seo"
version = "1.5.35"
description = "Sitemap XML generation helpers for Wagtail"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Gerry Mcbride" }]
dependencies = [
  "Django>=4.2",
  "wagtail>=5.2",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-django",
]
s3 = [
  "django-storages[s3]",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = ["src", "."]
testpaths = ["tests"]
